Move GhostGui.cs into UI folder and namespace
